Adds name of the won game to the notification.
It would be useful to see the name of the won game before clicking the notification to actually go to the won page. I tested this change to the notify() function and this just gets the name of the first game from won games list and add its to the notification message.

I know you said you wanted to "keep number of requests low to avoid temp bans.", but this request happens only when the game is already won which doesn't happen very often, since you only win a few games within a month. I also don't think a request to one html page on the site will get anyone temp banned, and this would be very useful piece of information to add.