added detailed remaining time
The span showing remaining time on each giveaway now shows a more detailed and accurate remaining time.
Two lines(199,200) of code added  to change the text of span and one function added at the bottom of file to chage epoc time to d:h:m:s and return a string to be placed inside span.